Fix crash on 1.14.4 due to incorrect signature
A crash was reported on the Fabric Discord server that I traced back to this method. The signature of the constructor has changed in 1.14.4 to include a BlockColors instance, so I've included that in the signature of the mixin method. It would probably be better to entirely remove the method since it's effectively empty, but I'll leave that to your disgression.